# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FanStage API",
    "version": "0.1.0",
    "description": "API documentation for FanStage backend. Update this file as endpoints are added."
  },
  "tags": [
    {
      "name": "Auth",
      "description": "User & authentication"
    }
  ],
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Register user",
        "tags": ["Auth"],
        "description": "Create a new user account. Returns token, status, and user.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Login (user or admin)",
        "tags": ["Auth"],
        "description": "Login with email and password. Returns token and user data.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/auth/register-admin": {
      "post": {
        "summary": "Register admin",
        "tags": ["Auth"],
        "description": "Create an admin account using a security code.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterAdminRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Invalid security code"
          }
        }
      }
    }
    ,
    "/api/auth/me": {
      "get": {
        "summary": "Get current user",
        "tags": ["Auth"],
        "description": "Return the authenticated user from the access token.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "user": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid token"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterUserRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "StrongPass123!"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "StrongPass123!"
          }
        }
      },
      "RegisterAdminRequest": {
        "type": "object",
        "required": ["name", "email", "password", "securityCode"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Admin User"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "StrongPass123!"
          },
          "securityCode": {
            "type": "string",
            "example": "YOUR_ADMIN_SECURITY_CODE"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user_123"
          },
          "name": {
            "type": "string",
            "example": "Jane Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "jane@example.com"
          },
          "role": {
            "type": "string",
            "enum": ["user", "admin"],
            "example": "user"
          },
          "accountType": {
            "type": "string",
            "enum": ["free", "premium"],
            "example": "free"
          },
          "accountStatus": {
            "type": "boolean",
            "example": true
          },
          "emailVerified": {
            "type": "boolean",
            "example": false
          },
          "profileComplete": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "status": {
            "type": "string",
            "example": "success"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}